// src/enums/routes.ts
export enum Routes {
    HomeLink = '/',
    CreditLink = '/credit',
    DepositLink = '/deposit',
    WalletLink = '/wallet',

    ProfileLink = '/profile',
    SignInLink = '/sign-in',
    YandexOauthLink = "/yandex/callback",

    History = '/history',
    HistoryCreditLink = '/history/credit',
    HistoryDepositLink = '/history/deposit',
    HistoryWalletLink = '/history/wallet',

}

export enum AuthRoutes{
    History = '/history',
    HistoryCreditLink = '/history/credit',
    HistoryDepositLink = '/history/deposit',
    HistoryWalletLink = '/history/wallet',
}